feat(casper): coin module api init with cleanup
