+++ .changeset/seven-suns-argue.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/coin-casper": minor
+"@ledgerhq/live-common": minor
+---
+
+coinModuleApi init for Casper with base folders structure cleanup
