+++ libs/coin-modules/coin-casper/src/bridge/broadcast.ts
@@ -2,7 +2,7 @@ import { patchOperationWithHash } from "@ledgerhq/ledger-wallet-framework/operat
 import { AccountBridge } from "@ledgerhq/types-live";
 import { Transaction as CasperTransaction, PublicKey } from "casper-js-sdk";
 import invariant from "invariant";
-import { broadcastTx } from "../api";
+import { broadcastTx } from "../network/api";
 import { Transaction } from "../types";
 
 export const broadcast: AccountBridge<Transaction>["broadcast"] = async ({
