+++ libs/coin-modules/coin-casper/src/bridge/deviceTransactionConfig.ts
@@ -3,7 +3,7 @@ import { log } from "@ledgerhq/logs";
 import { Account, AccountLike } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
 import { methodToString } from "../common-logic";
-import { CASPER_NETWORK } from "../consts";
+import { CASPER_NETWORK } from "../constants";
 import { Transaction, TransactionStatus } from "../types";
 
 export type ExtraDeviceTransactionField = {
