+++ libs/coin-modules/coin-casper/src/bridge/getTransactionStatus.ts
@@ -8,12 +8,13 @@ import {
 import { AccountBridge } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
 import {
+  CASPER_MAX_TRANSFER_ID,
   CASPER_MINIMUM_VALID_AMOUNT_MOTES,
   MayBlockAccountError,
   InvalidMinimumAmountError,
-} from "../consts";
+} from "../constants";
 import { CasperInvalidTransferId } from "../errors";
-import { CASPER_MAX_TRANSFER_ID, validateMemo } from "../logic/validateMemo";
+import { validateMemo } from "../logic/validateMemo";
 import { CasperAccount, Transaction, TransactionStatus } from "../types";
 import { getAddress, isAddressValid } from "./bridgeHelpers/addresses";
