+++ libs/coin-modules/coin-casper/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./common";
+export * from "./network";
 export * from "./signer";
