+++ libs/coin-modules/coin-casper/src/types/network.ts
